Removal redundant components
And fix css, and split resourceCard => ResourceCardWithUser and ResourceCarWithoutUser
