--- src/Styles/Components/Button/IconButtonStyles.tsx
@@ -1,14 +1,0 @@
-import { StyleSheet } from "react-native";
-
-export default StyleSheet.create({
-    container:{
-        alignItems: 'center',
-        justifyContent: 'center',
-        width: 40,
-        height: 40,
-        padding: 5,
-        borderRadius: 7,
-        transition: 0.2,
-        marginHorizontal: 15,
-    },
-});
