+++ src/Styles/Components/Button/LinkStyles.tsx
@@ -1,0 +1,12 @@
+import { StyleSheet } from "react-native";
+import { COLORS } from "../../Colors";
+
+export default StyleSheet.create({
+    btnBackground: {
+        width: '100%',
+    },
+    link: {
+        color: COLORS.AccentColor,
+        textAlign: 'center',
+    },
+});
