--- src/Styles/Components/Button/ReturnButtonStyles.tsx
@@ -1,15 +1,0 @@
-import { StyleSheet } from "react-native";
-
-export default StyleSheet.create({
-    returnBtn:{
-        alignItems: 'center',
-        justifyContent: 'center',
-        width: 40,
-        height: 40,
-        padding: 5,
-        borderRadius: 7,
-        transition: 0.2,
-        marginHorizontal: 15,
-        marginTop: 15,
-    },
-});
